Fix generation script to add implementation
Previously, the code meant to remove copyright blocks inadvertently
would delete entire files if they did not have an include guard.
Specifically, the code would search for the text `#ifndef __` and
replace everything up to its first occurrence. If that text was not
present, the `find` method would return `-1`, and the code would
delete all of `self.contents` except for the last character. The end
effect was that all implementation files, which do not have include
guards, would be deleted from the output.

This commit uses a different strategy to remove the copyright blocks. It
finds the first non-empty non-comment line and removes everything up to
there. This works for both header and implementation files.